get userdata after login + user & company store rework base
