+++ src/screens/Fairs/index.js
@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
 import Container from './container';
-import { actionCreators as userActions } from '../../redux/modules/user';
+import { actionCreators as companyActions } from '../../redux/modules/user';
 
 const mapStateToProps = (state) => {
-  const { user: { fairs } } = state;
+  const { company: { fairs } } = state;
 
   return {
     fairs
@@ -12,7 +12,7 @@ const mapStateToProps = (state) => {
 
 const mapDispatchToProps = dispatch => ({
   getFairs: () => {
-    dispatch(userActions.getFairs());
+    dispatch(companyActions.getFairs());
   }
 });
